[Refactor] Authentication 도메인 분리
