+++ app/src/main/java/fc/be/app/common/authentication/controller/dto/response/TokenResponse.java
@@ -1,0 +1,6 @@
+package fc.be.app.common.authentication.controller.dto.response;
+
+public record TokenResponse(
+        String token
+) {
+}
