--- app/src/main/java/fc/be/app/domain/member/dto/response/TokenResponse.java
@@ -1,6 +1,0 @@
-package fc.be.app.domain.member.dto.response;
-
-public record TokenResponse(
-        String token
-) {
-}
