+++ app/src/main/java/fc/be/app/global/config/security/filter/LoginAuthenticationFilter.java
@@ -1,7 +1,7 @@
 package fc.be.app.global.config.security.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import fc.be.app.domain.member.dto.request.LoginRequest;
+import fc.be.app.common.authentication.controller.dto.request.LoginRequest;
 import fc.be.app.global.config.security.model.token.LoginAuthenticationToken;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
